Add some documentation and address PR concerns
