+++ CreeDictionary/search_quality/__init__.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from pathlib import Path
 from typing import TypedDict
