+++ CreeDictionary/search_quality/sample.csv
@@ -1,14 +1,15 @@
-English,Nêhiyawêwin 1,Nêhiyawêwin 2,Nêhiyawêwin 3,,,Poor result,Notes
+Query,Nêhiyawêwin 1,Nêhiyawêwin 2,Nêhiyawêwin 3,,,Poor result,Notes
 star,acâhkos,atâhk,,,,,
 see,wâpamêw,wâpiw,wâpahtam,,,kwêyask,"Ignore ""\[see"""
 eat,mîcisow,mîciw,mowêw,,,,
 fire,iskotêw,,,,,,
 happy birthday,miyo-tipiska,,,,,,
-thank you,ay-hay,hay-hay,,,,,
-dog,atim,,,,,,very irregular inflections
-horse,misatim,,,,,,
-'horse',misatim,,,,,,
-"""horse""",misatim,,,,,,
+thank you,ay-hay,hay-hay,kinanâskomitin,,,,
+welcome,tatawâw,,,,,,
+dog,atim,nitêm,,,,,very irregular inflections
+horse,misatim,mistatim,,,,,
+'horse',misatim,mistatim,,,,,
+"""horse""",misatim,mistatim,,,,,
 beaverberry,amiskomin,amiskominak,,,,,
 beaverberries,amiskomin,amiskominak,,,,,
 beaver-berry,amiskomin,amiskominak,,,,,
@@ -35,7 +36,7 @@ goodbye,ekosi mâka ,,,,,,
 work together,mâmawi-isîhcikêwin,,,,,,
 working together,mâmawi-isîhcikêwin,,,,,,
 work togeth,mâmawi-isîhcikêwin,,,,,,
-Cree,nêhiyaw,,,,,,
+Cree,nêhiyaw,nêhiyawêwin,,,,,
 Edmonton,amiskwaciy-wâskahikan,,,,,,
 edmonton,amiskwaciy-wâskahikan,,,,,,
 beaver,amisk,,,,,,
@@ -45,17 +46,22 @@ beaver hill,natimîwiyiniwak,amiskwacîwiyiniwak,,,,,
 beaver hills,natimîwiyiniwak,amiskwacîwiyiniwak,,,,,
 Calgary,otôskwanihk,,,,,,
 calgary,otôskwanihk,,,,,,
-cat,minos,,,,,,
-car,pôsîw,pôsiy,,,,,
+car,sêhkêpayîs,,,,,,
 rabbit,wâpos,,,,,,
 water,nipiy,,,,,,
 telling a story,âcimêw,,,,,,
-on water,nipîhk,,,,,,
+on water,nipîhk,,,,,,"aspirational search, as result is based on inflections, not a direct definition entry"
 hello,tânisi,,,,,,
 difficult thing,âyiman,,,,,,
 many snakes,kinêpikoskâw,,,,,,
-dances,nîmiw,,,,,,
-i see you,kinêpikoskâw,,,,,,
+there are many snakes,kinêpikoskâw,,,,,,
+dances,nîmiw,nîmihitow,,,,,
+dancing time,nîminâniwan,nîmiw,,,,,
+telling a story,âcimow,,,,,,
+go to bed,mâci-nipa,,,,,,
+Merry Christmas,miyo-manitowi-kîsikânisik,,,,,,
+merry christmas,miyo-manitowi-kîsikânisik,,,,,,
+i see you,kiwâpamitin,,,,,,
 he sees him,wâpamêw,,,,,,
 she sees him,wâpamêw,,,,,,
 starblanket,atâhkakohp,acâhkosa kâ-otakohpit,,,,,
